services/airex-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airex-core"
version = "0.1.0"
description = "AIREX shared core package"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.0,<3.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "sqlalchemy[asyncio]>=2.0.37,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "alembic>=1.14.0,<2.0.0",
    "pgvector>=0.3.5,<1.0.0",
    "redis[hiredis]>=5.2.0,<6.0.0",
    "arq>=0.26.1,<1.0.0",
    "litellm>=1.60.0,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt>=4.0.0,<4.1.0",
    "email-validator>=2.1.0,<3.0.0",
    "boto3>=1.34.0",
    "asyncssh>=2.14.0",
    "google-cloud-logging>=3.10.0",
    "google-cloud-compute>=1.15.0",
    "google-auth>=2.25.0",
    "httpx>=0.28.0,<1.0.0",
    "structlog>=24.4.0,<26.0.0",
    "prometheus-client>=0.21.0,<3.0.0",
    "pyyaml>=6.0",
]

[tool.setuptools]
packages = ["airex_core"]
